Comment out debug logging in FrameReceiver
