+++ CMakeLists.txt
@@ -32,9 +32,10 @@ if(PLATFORM_DESKTOP)
     target_include_directories(MRDesktopServer PRIVATE ${COMMON_INCLUDES})
 
     # Create console client executable
-    add_executable(MRDesktopConsoleClient 
+    add_executable(MRDesktopConsoleClient
         src/clients/console/main.cpp
         src/shared/FrameLogger.cpp
+        src/shared/FrameReceiver.cpp
     )
     target_include_directories(MRDesktopConsoleClient PRIVATE ${COMMON_INCLUDES})
 
@@ -59,7 +60,7 @@ if(PLATFORM_ANDROID OR BUILD_ANDROID_CLIENT)
     # Android native client library
     add_library(MRDesktopAndroidClient SHARED
         android/app/src/main/cpp/native_client.cpp
-        android/app/src/main/cpp/frame_receiver.cpp
+        src/shared/FrameReceiver.cpp
         android/app/src/main/cpp/input_sender.cpp
     )
