+++ android/app/src/main/cpp/native_client.cpp
@@ -2,7 +2,7 @@
 #include <android/log.h>
 #include <android/native_window.h>
 #include <android/native_window_jni.h>
-#include "frame_receiver.h"
+#include "FrameReceiver.h"
 #include "input_sender.h"
 #include <thread>
 #include <atomic>
