+++ src/clients/console/main.cpp
@@ -10,6 +10,7 @@
 #include <algorithm>
 #include "protocol.h"
 #include "../shared/FrameLogger.h"
+#include "../shared/FrameReceiver.h"
 
 #pragma comment(lib, "ws2_32.lib")
 
@@ -57,209 +58,8 @@ public:
     }
 };
 
-// Helper function to dump hex data for debugging
-void HexDump(const char* data, size_t size, const std::string& label) {
-    std::cout << label << " (size=" << size << "):" << std::endl;
-    size_t maxBytes = (size < 32) ? size : 32;
-    for (size_t i = 0; i < maxBytes; i++) {
-        printf("%02X ", (unsigned char)data[i]);
-        if ((i + 1) % 16 == 0) std::cout << std::endl;
-    }
-    if (size > 32) std::cout << "... (truncated)";
-    std::cout << std::endl;
-}
 
-class FrameReceiver {
-private:
-    SOCKET m_Socket = INVALID_SOCKET;
-    std::vector<BYTE> m_FrameBuffer;
-    
-public:
-    bool Connect(const char* serverIP, int port) {
-        m_Socket = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
-        if (m_Socket == INVALID_SOCKET) {
-            std::cerr << "Failed to create socket: " << WSAGetLastError() << std::endl;
-            return false;
-        }
-        
-        sockaddr_in serverAddr;
-        serverAddr.sin_family = AF_INET;
-        serverAddr.sin_port = htons(port);
-        
-        if (inet_pton(AF_INET, serverIP, &serverAddr.sin_addr) <= 0) {
-            std::cerr << "Invalid server IP address" << std::endl;
-            closesocket(m_Socket);
-            m_Socket = INVALID_SOCKET;
-            return false;
-        }
-        
-        if (connect(m_Socket, (sockaddr*)&serverAddr, sizeof(serverAddr)) == SOCKET_ERROR) {
-            std::cerr << "Failed to connect to server: " << WSAGetLastError() << std::endl;
-            closesocket(m_Socket);
-            m_Socket = INVALID_SOCKET;
-            return false;
-        }
-        
-        // Set socket to non-blocking mode for frame receiving
-        u_long mode = 1;
-        if (ioctlsocket(m_Socket, FIONBIO, &mode) != 0) {
-            std::cerr << "Failed to set socket non-blocking: " << WSAGetLastError() << std::endl;
-            closesocket(m_Socket);
-            m_Socket = INVALID_SOCKET;
-            return false;
-        }
-        
-        std::cout << "Connected to server successfully!" << std::endl;
-        return true;
-    }
-    
-    bool ReceiveFrame(FrameMessage& frameMsg, std::vector<BYTE>& frameData, int frameNumber) {
-        if (m_Socket == INVALID_SOCKET) return false;
-        
-        std::cout << "CLIENT RECV: Frame " << frameNumber << " - Expecting header size: " << sizeof(FrameMessage) << std::endl;
-        
-        // Receive frame message header
-        int received = recv(m_Socket, (char*)&frameMsg, sizeof(FrameMessage), 0);
-        
-        if (received == SOCKET_ERROR) {
-            int error = WSAGetLastError();
-            if (error == WSAEWOULDBLOCK) {
-                // No data available right now, not an error
-                return false;
-            }
-            std::cout << "CLIENT ERROR: Frame " << frameNumber << " - Socket error on header recv: " << error << std::endl;
-            return false;
-        }
-        
-        std::cout << "CLIENT RECV: Frame " << frameNumber << " - Header received: " << received << " bytes" << std::endl;
-        
-        // Debug: Show raw header bytes
-        HexDump((char*)&frameMsg, received, "CLIENT RECV: Raw header data");
-        
-        if (received != sizeof(FrameMessage)) {
-            if (received == 0) {
-                std::cout << "Server disconnected" << std::endl;
-            } else if (received > 0) {
-                std::cout << "CLIENT ERROR: Frame " << frameNumber << " - Partial header! Expected: " 
-                         << sizeof(FrameMessage) << ", Got: " << received << std::endl;
-                std::cout << "ABORTING on first bad frame for debugging" << std::endl;
-                exit(1);
-            }
-            return false;
-        }
-        
-        std::cout << "CLIENT RECV: Frame " << frameNumber << " - Header parsed: type=" << frameMsg.header.type 
-                 << " (hex: 0x" << std::hex << frameMsg.header.type << std::dec << ")"
-                 << ", headerSize=" << frameMsg.header.size
-                 << ", w=" << frameMsg.width << ", h=" << frameMsg.height 
-                 << ", dataSize=" << frameMsg.dataSize << std::endl;
-                 
-        // Debug: Show structure layout info
-        std::cout << "CLIENT DEBUG: sizeof(FrameMessage)=" << sizeof(FrameMessage) 
-                 << ", sizeof(MessageHeader)=" << sizeof(MessageHeader) << std::endl;
-        
-        // Verify this is a frame message
-        if (frameMsg.header.type != MSG_FRAME_DATA) {
-            std::cout << "CLIENT ERROR: Frame " << frameNumber << " - Unexpected message type: " 
-                     << frameMsg.header.type << std::endl;
-            std::cout << "ABORTING on first bad frame for debugging" << std::endl;
-            exit(1);
-        }
-        
-        // Validate frame dimensions are reasonable
-        if (frameMsg.width == 0 || frameMsg.height == 0 ||
-            frameMsg.width > 10000 || frameMsg.height > 10000 ||
-            frameMsg.dataSize > 100000000) {
-            std::cout << "CLIENT ERROR: Frame " << frameNumber << " - Invalid dimensions: w=" 
-                     << frameMsg.width << ", h=" << frameMsg.height << ", dataSize=" << frameMsg.dataSize << std::endl;
-            std::cout << "ABORTING on first bad frame for debugging" << std::endl;
-            exit(1);
-        }
-        
-        // Resize buffer if needed
-        if (m_FrameBuffer.size() < frameMsg.dataSize) {
-            m_FrameBuffer.resize(frameMsg.dataSize);
-        }
-        
-        std::cout << "CLIENT RECV: Frame " << frameNumber << " - Expecting pixel data: " << frameMsg.dataSize << " bytes" << std::endl;
-        
-        // Receive frame pixel data (must receive ALL data for this frame)
-        UINT32 totalReceived = 0;
-        while (totalReceived < frameMsg.dataSize) {
-            received = recv(m_Socket, (char*)m_FrameBuffer.data() + totalReceived, 
-                           frameMsg.dataSize - totalReceived, 0);
-            
-            if (received == SOCKET_ERROR) {
-                int error = WSAGetLastError();
-                if (error == WSAEWOULDBLOCK) {
-                    // No data available, wait a bit and try again
-                    std::this_thread::sleep_for(std::chrono::milliseconds(1));
-                    continue;
-                }
-                std::cout << "CLIENT ERROR: Frame " << frameNumber << " - Socket error receiving pixel data at offset " 
-                         << totalReceived << "/" << frameMsg.dataSize << ": " << error << std::endl;
-                std::cout << "ABORTING on first bad frame for debugging" << std::endl;
-                exit(1);
-            }
-            
-            if (received == 0) {
-                std::cout << "CLIENT ERROR: Frame " << frameNumber << " - Server disconnected during pixel data at offset " 
-                         << totalReceived << "/" << frameMsg.dataSize << std::endl;
-                std::cout << "ABORTING on first bad frame for debugging" << std::endl;
-                exit(1);
-            }
-            
-            totalReceived += received;
-        }
-        
-        std::cout << "CLIENT RECV: Frame " << frameNumber << " - Pixel data received: " << totalReceived << " bytes - COMPLETE" << std::endl;
-        
-        // Debug: Show last few bytes of pixel data to check for corruption
-        if (totalReceived >= 16) {
-            HexDump((char*)m_FrameBuffer.data() + totalReceived - 16, 16, "CLIENT RECV: Last 16 bytes of pixel data");
-        }
-        
-        // Copy frame data
-        frameData.clear();
-        frameData.resize(frameMsg.dataSize);
-        memcpy(frameData.data(), m_FrameBuffer.data(), frameMsg.dataSize);
-        
-        // Debug: Check if there are any pending bytes in the socket buffer
-        u_long bytesAvailable = 0;
-        if (ioctlsocket(m_Socket, FIONREAD, &bytesAvailable) == 0) {
-            std::cout << "CLIENT DEBUG: Frame " << frameNumber << " - Bytes still in socket buffer: " << bytesAvailable << std::endl;
-            
-            if (bytesAvailable > 0) {
-                std::cout << "CLIENT WARNING: Frame " << frameNumber << " - Unexpected extra data in buffer!" << std::endl;
-                
-                // Peek at the extra data
-                char peekBuffer[32];
-                u_long peekSize = (bytesAvailable < 32) ? bytesAvailable : 32;
-                int peeked = recv(m_Socket, peekBuffer, peekSize, MSG_PEEK);
-                if (peeked > 0) {
-                    HexDump(peekBuffer, peeked, "CLIENT DEBUG: Extra buffer data");
-                }
-            }
-        }
-        
-        return true;
-    }
-    
-    void Disconnect() {
-        if (m_Socket != INVALID_SOCKET) {
-            closesocket(m_Socket);
-            m_Socket = INVALID_SOCKET;
-        }
-    }
-    
-    SOCKET GetSocket() const { return m_Socket; }
-    
-    ~FrameReceiver() {
-        Disconnect();
-    }
-};
-
-void SaveFrameAsBMP(const FrameMessage& frameMsg, const std::vector<BYTE>& frameData, const std::string& filename) {
+void SaveFrameAsBMP(const FrameMessage& frameMsg, const std::vector<uint8_t>& frameData, const std::string& filename) {
     // Simple BMP save for debugging (assumes BGRA format)
     BITMAPFILEHEADER fileHeader = {};
     BITMAPINFOHEADER infoHeader = {};
@@ -387,7 +187,7 @@ int main(int argc, char* argv[]) {
     
     // Variables for frame receiving and input handling
     FrameMessage frameMsg;
-    std::vector<BYTE> frameData;
+    std::vector<uint8_t> frameData;
     int frameCount = 0;
     bool savedFirstFrame = false;
     auto startTime = std::chrono::high_resolution_clock::now();
